fix: kvcache offloading
