+++ tests/model_serving/model_server/llmd/conftest.py
@@ -529,6 +529,10 @@ def _create_llmisvc_from_config(
     if service_account:
         template["serviceAccountName"] = service_account
 
+    volumes = config_cls.template_volumes()
+    if volumes:
+        template["volumes"] = volumes
+
     prefill = config_cls.prefill_config()
     if prefill and service_account and "template" in prefill:
         prefill["template"]["serviceAccountName"] = service_account
@@ -548,12 +552,10 @@ def _create_llmisvc_from_config(
         "prefill": prefill,
         "worker": config_cls.worker_config(),
         "parallelism": config_cls.parallelism_config(),
+        "kv_cache_offloading": config_cls.kv_cache_offloading(),
     }
 
     LOGGER.info(f"\n{config_cls.format_describe(namespace=namespace)}")
-
-    if config_cls.kv_cache_offloading is not None:
-        svc_kwargs["kv_cache_offloading"] = config_cls.kv_cache_offloading
 
     with LLMInferenceService(**svc_kwargs) as llm_service:
         wait_for_llmisvc(llmisvc=llm_service, timeout=config_cls.wait_timeout)
