+++ tests/model_serving/model_server/llmd/llmd_configs/config_base.py
@@ -89,7 +89,13 @@ class LLMISvcConfig:
     def prefill_config(cls):
         return None
 
-    kv_cache_offloading = None
+    @classmethod
+    def kv_cache_offloading(cls):
+        return None
+
+    @classmethod
+    def template_volumes(cls):
+        return None
 
     @classmethod
     def worker_config(cls):
